moved protected route to src
